ai-playground-backend/pyproject.toml
[project]
name = "ai-playground-backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sarvamai>=0.1.22",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
